requirements.txt
bottle==0.12.15
requests==2.20.1
validate_email==1.3
simplejson==3.13.2
pycrypto==2.6.1
pystache==0.5.4
git+git://github.com/orseni/pydbwrapper@v1.1.3
python-jose==3.0.1
python-dotenv
